fix: only authors can edit and delete their profiles and photos

+++ Django-Basics/Django-basics-projects/petstagram/petstagram/photos/views.py
@@ -1,6 +1,6 @@
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.shortcuts import render, redirect
+from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
+from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse_lazy
 from django.views.generic import CreateView, UpdateView, DetailView
 
@@ -21,13 +21,29 @@ class PhotoAddPage(LoginRequiredMixin, CreateView):
         return super().form_valid(form)
 
 
-class PhotoEditPage(LoginRequiredMixin, UpdateView):
+class PhotoEditPage(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
     model = Photo
     form_class = PhotoEditForm
-    template_name = 'photos/photo-edit-page.html'
+    template_name = "photos/photo-edit-page.html"
+
+    def test_func(self):
+        photo = get_object_or_404(Photo, pk=self.kwargs["pk"])
+        return self.request.user == photo.user
 
     def get_success_url(self):
-        return reverse_lazy('photo-details', kwargs={'pk': self.object.pk})
+        return reverse_lazy("photo-details", kwargs={"pk": self.object.pk})
+
+
+# class PhotoDetailsView(LoginRequiredMixin, DetailView):
+#     model = Photo
+#     template_name = 'photos/photo-details-page.html'
+#
+#     def get_context_data(self, **kwargs):
+#         context = super().get_context_data(**kwargs)
+#         context['likes'] = self.object.photo_likes.all()
+#         context['comments'] = self.object.photo_comments.all()
+#         context['comment_form'] = CommentForm()
+#         return context
 
 
 class PhotoDetailsView(LoginRequiredMixin, DetailView):
@@ -36,16 +52,21 @@ class PhotoDetailsView(LoginRequiredMixin, DetailView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
+
         context['likes'] = self.object.photo_likes.all()
         context['comments'] = self.object.photo_comments.all()
         context['comment_form'] = CommentForm()
+        self.object.has_liked = self.object.photo_likes.filter(user=self.request.user).exists()
+
         return context
 
 
 @login_required
 def photo_delete(request, pk):
     photo = Photo.objects.get(pk=pk)
-    photo.delete()
+
+    if request.user == photo.user:
+        photo.delete()
 
     return redirect("home")
